Refactor code move content of old ‘core.student’ module in ‘core.user’
