+++ linnote/client/admin.py
@@ -13,8 +13,7 @@ from flask import redirect, render_template, request, url_for
 from flask_login import login_required
 from linnote.core.assessment import Assessment
 from linnote.core.report import Report
-from linnote.core.student import Group
-from linnote.core.user import User
+from linnote.core.user import User, Group
 from .forms import ReportForm, GroupForm, UserForm
 from .utils import session
